issue/requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.3.1
attrs==20.3.0
autopep8==1.5.6
aws-psycopg2==1.2.1
awsebcli==3.19.3
bcrypt==3.1.7
beautifulsoup4==4.8.2
blessed==1.17.12
blinker==1.4
boto3==1.17.3
botocore==1.20.3
Brlapi==0.7.0
cached-property==1.5.2
cairocffi==1.2.0
CairoSVG==2.5.1
cement==2.8.2
certifi==2020.12.5
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cssselect2==0.4.1
cupshelpers==1.0
cx-Oracle==8.1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.6
django-ckeditor==6.0.0
django-countries==7.0
django-crispy-forms==1.11.0
django-js-asset==1.2.2
django-storages==1.11.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
docker==4.4.1
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.16.0
html5lib==1.1
httplib2==0.14.0
idna==2.10
jmespath==0.10.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pathspec==0.5.9
pexpect==4.6.0
Pillow==8.1.0
ply==3.11
protobuf==3.6.1
pycairo==1.16.2
pycity==0.1.3
pycodestyle==2.7.0
pycparser==2.19
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
Pyphen==0.10.0
pyRFC3339==1.1
pyrsistent==0.17.3
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
pytz==2021.1
pyuca==1.2
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.24.0
requests-unixsocket==0.2.0
s3transfer==0.3.4
SecretStorage==2.3.1
semantic-version==2.5.0
simplejson==3.16.0
six==1.15.0
soupsieve==1.9.5
sqlparse==0.4.1
systemd-python==234
termcolor==1.1.0
texttable==1.6.3
tinycss2==1.1.0
toml==0.10.2
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.11
usb-creator==0.3.7
virtualenv==20.4.2
wadllib==1.3.3
wcwidth==0.1.9
WeasyPrint==52.2
webencodings==0.5.1
websocket-client==0.57.0
xcffib==0.8.1
xkit==0.0.0
